renderer-view/native: Allow synchronizing frame clock updates with actor updates
Allow per-view sychnronization of frame clock updates to updates
generated by a specified actor.

This uses the variable mode of ClutterFrameClock to either perform an
immediate update or a delayed update depending on whether the update
is performed in the context of the specified actor or a different
actor.

This approach allows minimum latency between updates of the actor and
updates of the view, and avoids updates of the view when the
specified actor has not updated.

If the specified actor is frozen or destroyed, the synchornization is
stopped as no further updates from the actor are expected.