Core: Include a small subset of xcursor in-tree
For a Wayland only build, we would like to avoid linking against libXucrsor
which on it turn, links back to all the X11.
In order to achieve that, we include a small subset of xcursor that
is needed for a Wayland only  build cursor loading.

In case Mutter is built with X11 or with both Wayland & X11, we link against
libXcursor and don't make use of the in-tree implementation.

This patch mimics what GTK 4 do by shipping an in-tree copy of xcursor.
Especially that libwayland-cursor doesn't provide an alternative to xcursor
itself.